updates
updating the photo of all the jpb history and proprly formating it. Had to go into CSS to make all the images the same width and height to look clean
had to switch different logos to logos that fit the screen better. Stretching the logos is not always the answer, so  went in to alter logos that were around the same format and edited the sizing